LayoutTests/imported/w3c/web-platform-tests/imagebitmap-renderingcontext/bitmaprenderer-as-imagesource-expected.txt


FAIL Test that createImageBitmap from a bitmaprenderer canvas produces correct result assert_equals: expected 10 but got 20
PASS Test that createImageBitmap on a bitmaprenderer canvas that never consumes any source produces correct result 
PASS Test that createImageBitmap on a bitmaprenderer canvas that consumes null produces correct result 
